WIP: Add meson formatting
Meson has a recommended formatting style[1].  Muon is a C99
implementation of meson that implements an fmt command that enforces
this style.  While we ideally would like to use the default meson build
for this some day, we work with what we have.

Muon fmt is used for the vscode-meson plugin[2] so it should remain
supported for a while given the ubiquity of that plugins use.

[1] https://mesonbuild.com/Style-guide.html
[2] https://github.com/mesonbuild/vscode-meson/blob/10eb43d30a6d74c841abefca2597a33a900cf77e/src/tools/muon.ts#L45

Signed-off-by: Ed Tanous <ed@tanous.net>
Change-Id: I88c2631de1e258bb7874188f3b77ff58f38786bf